clubmason/problems/serializers/contentserializer.py
from rest_framework import serializers
from problems.models.content import Content

class ContentSerializer(serializers.ModelSerializer):
    class Meta:
        model = Content
        fields = '__all__'