clubmason/requirements.txt
django
djangorestframework
django-taggit

